midlewares de autenticação e regra de negocio renovada

+++ avaliacao-final/app/controllers/categorias_controller.ts
@@ -3,42 +3,49 @@ import Categoria from '#models/categoria'
 import { createCategoriaValidator } from '#validators/categoria'
 
 export default class CategoriasController {
-  // Listar todas as categorias
+  // PÚBLICO: Listar todas as categorias
   async index() {
-    const categorias = await Categoria.all()
-    return categorias
+    return await Categoria.all()
   }
 
-  // Cadastrar nova categoria
-  async store({ request, response }: HttpContext) {
+  // PROTEGIDO: Cadastrar nova categoria
+  async store({ request, auth, response }: HttpContext) {
+    if (auth.user?.role !== 'gerente') {
+      return response.forbidden({ message: 'Apenas gerentes podem cadastrar categorias.' })
+    }
+
     const payload = await request.validateUsing(createCategoriaValidator)
     const categoria = await Categoria.create(payload)
-
     return response.created(categoria)
   }
 
-  // Buscar uma única categoria pelo ID
+  // PÚBLICO: Buscar uma única categoria pelo ID
   async show({ params }: HttpContext) {
-    const categoria = await Categoria.findOrFail(params.id)
-    return categoria
+    return await Categoria.findOrFail(params.id)
   }
 
-  // Atualizar uma categoria
-  async update({ params, request }: HttpContext) {
+  // PROTEGIDO: Atualizar uma categoria
+  async update({ params, request, auth, response }: HttpContext) {
+    if (auth.user?.role !== 'gerente') {
+      return response.forbidden({ message: 'Apenas gerentes podem atualizar categorias.' })
+    }
+
     const categoria = await Categoria.findOrFail(params.id)
     const payload = await request.validateUsing(createCategoriaValidator)
 
     categoria.merge(payload)
     await categoria.save()
-
     return categoria
   }
 
-  // Excluir uma categoria
-  async destroy({ params, response }: HttpContext) {
+  // PROTEGIDO: Excluir uma categoria
+  async destroy({ params, auth, response }: HttpContext) {
+    if (auth.user?.role !== 'gerente') {
+      return response.forbidden({ message: 'Apenas gerentes podem deletar categorias.' })
+    }
+
     const categoria = await Categoria.findOrFail(params.id)
     await categoria.delete()
-
     return response.noContent()
   }
 }
